Fix border resistance not coherent on modes

+++ LittleBigMouse.Core/LittleBigMouse.Zones/ZoneLink.cs
@@ -11,7 +11,7 @@ public class ZoneLink : IZonesSerializable
     public int SourceToPixel { get; set; }
     public int TargetFromPixel { get; set; }
     public int TargetToPixel { get; set; }
-    public int BorderResistance { get; set; }
+    public double BorderResistance { get; set; }
     [JsonIgnore] public Zone? Target { get; set; }
     public int TargetId => Target?.Id??-1;
