+++ LittleBigMouse.Hook/Engine/ZonesLayout.cpp
@@ -128,7 +128,7 @@ ZoneLink* GetNewZoneLink(tinyxml2::XMLElement* element)
             const long targetFromPixel = XmlHelper::GetLong(zoneElement,"TargetFromPixel");
             const long targetToPixel = XmlHelper::GetLong(zoneElement,"TargetToPixel");
             const long targetId = XmlHelper::GetLong(zoneElement,"TargetId");
-            const long borderResistance = XmlHelper::GetLong(zoneElement,"BorderResistance");
+            const double borderResistance = XmlHelper::GetDouble(zoneElement,"BorderResistance");
 
 			*currentLink = new ZoneLink(from,to,sourceFromPixel,sourceToPixel,targetFromPixel,targetToPixel,borderResistance,targetId);
